+++ src/main/java/Group/Better/repository/PostRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface PostRepository {
 
     @Insert("insert into posts (title, content, user_id) values (#{title}, #{content}, #{user_id})")
-    void insert(@Param("title") String title, @Param("content") String content, @Param("user_id") long user_id);
+    void insert(@Param("title") String title, @Param("content") String content, @Param("user_id") int user_id);
 
     @Select("select * from posts")
     List<Post> findAll();
